requirements.txt

albumentations==1.1.0
altair==4.2.0
coremltools==5.2.0
Flask==2.1.2
ipython==8.3.0
matplotlib==3.5.1
numpy==1.22.3
onnx==1.11.0
onnxruntime==1.11.0
opencv_python==4.5.5.64
openvino==2022.1.0
pafy==0.5.5
pandas==1.3.5
Pillow==9.1.0
protobuf==3.20.1
psutil==5.9.0
pycocotools==2.0.4
PyMySQL==1.0.2
PyYAML==6.0
requests==2.27.1
s3fs==2022.3.0
scipy==1.8.0
seaborn==0.11.2
setuptools==58.1.0
streamlit==1.8.1
tensorflow==2.9.0
tensorflowjs==3.17.0

tflite_runtime==2.8.0
thop==0.0.31.post2005241907
torch==1.11.0
torchvision==0.12.0
tqdm==4.63.1
